requirements.txt
streamlit
tensorflow
numpy
matplotlib
pandas
Pillow